# Fix perf regression in System.Net.NetworkInformation.PhysicalAddress