Closes #39720 by:
* Using the internal API directly to avoid input validation overhead
* Encourage inlining of a method that was in-line prior to #39702


|  Method |        Job | Toolchain |      Mean |    Error |   StdDev | Ratio | RatioSD |
|-------- |----------- |---------- |----------:|---------:|---------:|------:|--------:|
| PAShort | Job-CSJLAA |        PR |  11.90 ns | 0.215 ns | 0.191 ns |  0.84 |    0.02 |
| PAShort | Job-RPOJLS |    master |  14.17 ns | 0.302 ns | 0.335 ns |  1.00 |    0.00 |

/cc @danmosemsft @DrewScoggins