@stephentoub 2020-08-09
Is this addressing a regression from 3.1 or just intra-5.0?

While I appreciate the effort here, I'm inclined to just accept the small regression and close this, especially if it's no worse than 3.1. The aggressive inlining will affect all call sites of that internal helper, this internal helper wasn't otherwise being used by this assembly and thus will add to places it's included in shipped binaries, and I never like using internal instead of public APIs when there's a public API that is meant to suffice. On top of that, I'm skeptical there are places where the cited 3ns in Physical address.ToString will make a real difference (but I'd like to better understand such scenarios if there is one).

Thanks.