@geoffkizer 2020-08-09
If we are really concerned about the perf here, then it seems like we should just take the perf PR you mentioned (#39702), which would benefit everyone using this new API.