@tkp1n 2020-08-09
There are actually two regressions if we look at the entire time window since 3.1:

1. With 3.1 PhysicalAddress.ToString had custom hex encoding logic. This was replaced by the shared but slower branch-free encoding method in HexConverter.
2. The introduction of the public API caused a method to longer be inlined and added some overhead due to input validation.

|  Method |        Job |       Runtime | Toolchain |      Mean |     Error |    StdDev | Ratio | RatioSD |
|-------- |----------- |-------------- |---------- |----------:|----------:|----------:|------:|--------:|
| PAShort |        3.1 | .NET Core 3.1 |   Default |  9.290 ns | 0.2043 ns | 0.2187 ns |  0.65 |    0.02 |
| PAShort | Job-YOCJAS | .NET Core 5.0 |        PR | 12.163 ns | 0.1314 ns | 0.1229 ns |  0.86 |    0.01 |
| PAShort | Job-MOXHEH | .NET Core 5.0 |    master | 14.203 ns | 0.1427 ns | 0.1265 ns |  1.00 |    0.00 |

As the combined regression since 3.1 is quite substantial I would suggest the following:
1. We add the `AggressiveInlining` with this PR to restore how things were before #37546 but stick to using the public API (accepting its slight overhead).
2. Focus on getting improvements (either the vectorization proposed in #39702 or just plain faster scalar code) made to HexConverter to bring faster perf for every caller of the internal and public APIs.