@stephentoub 2020-08-09
Is the regression only on "PAShort"?  That's a one-byte PhysicalAddress.  Where does the throughput of using PhysicalAddress.ToString with a one-byte address show up as meaningful?