@tkp1n 2020-08-09
I would argue that neither `PAShort` nor `PAShort` are particularly relevant in the real world, assuming the typical usage would be 6-byte MAC addresses.
`PALong` seems to be bound by GC with high jitter on my machine and no regression compared to 3.1 outside of margin of error.