@stephentoub 2020-08-09
Is there an example you believe is real-world, does that exhibit a meaningful regression, and can we reasonably construct a use-case where such a regression matters? That's what I'm struggling with. In what situation is someone using PhysicalAddress.ToString on a hot path, and in such a use case, do we actually see a measurable regression?