@tkp1n 2020-08-10
I can't come up with a reason why anyone would use this on the hot path... but then again, I assume this benchmark was added to dotnet/performance for a reason.
Should it end up on the hot path of some application, I would assume the inherent that GC pressure caused by the created strings is worse issue than this 3ns regression.