@stephentoub 2020-08-10
> but then again, I assume this benchmark was added to dotnet/performance for a reason

I believe I added them since some work was being done to improve the performance by a community member in a previous release, and perf tests are needed to help validate and lock in such efforts. But just because a small regression in a perf test is identified, doesn't mean it needs to be fixed... it just means the test alerted us to something, at which point we can decide whether it's meaningful. My opinion on this would be different if, for example, the regression was 10x rather than 3ns. Until we can come up with a real scenario where the 3ns matters, is prefer to keep the code simple, using the same public APIs we ask everyone else to use, keeping assembly size smaller, and not risking regressions in the other places that consume the method being attributed as aggressive inlining.