@danmosemsft 2020-08-10
OK, seems we have consensus to leave the code. @tkp1n apologies for leading you down this dead end.